requirements.txt
chromadb
sentence-transformers
bs4
pdfplumber
openai
google-generativeai
python-dotenv